"12:40pm. Done with breakfast.
12:55pm. Let me finish my reading and I will resume?

Doing what exactly? Well, I just browse the AWS site for a bit. Right now my focus is extremely low, I really need to get some practical experience to digest the book learning. But until I get access to my financials sorted out, it can't be helped.

I'll do a tour of AWS and Azure, track down that ML related cloud service. It is going to be a bit tedious, maybe I'll take a break earlier today than usual. Who knows how long this will take.

If I let the current crop of applications pass and apply again in a few months instead that is fine.

1:20pm. I've started. My tension is uselessly high right now.

https://aws.amazon.com/getting-started/guides/deploy-webapp-ec2/module-one/

This is too rough to just read. It too boring unless I can follow along.

Let me change gears and check out Azure.

https://docs.microsoft.com/en-us/learn/azure/

https://customers.microsoft.com/en-us/story/1498781140435260527-forza-horizon-5-crosses-finish-line-fueled-by-azure-kubernetes-service
> When it launched, Forza Horizon 5 saw more than 10 million concurrent players—the biggest first week in Xbox Game Studios history—drive across vibrant, virtual landscapes and face off in thrilling tournaments and challenges. Forza Horizon 5 is also a stunning win behind the scenes, where Azure platform as a service (PaaS) provides a virtual crew, and a Windows-based container architecture adapts to changing demand at warp speed.

> "With our compute, storage, and data stacks all on Azure, our engineering teams can now spend more time building new experiences for our players rather than managing our infrastructure."

https://youtu.be/2vMEQ5zs1ko
Kubernetes vs. Docker: It's Not an Either/Or Question

https://azure.microsoft.com/en-us/free/

Free Azure account for 30 days. Maybe this thing won't ask me for my credit card.

> The answer came from a technology that ForzaTech didn’t expect—Kubernetes, the popular container orchestration platform. “We had never used container technology whatsoever. There was never any reason to even look into Kubernetes,” Hennessy recalls. But the team leaned into the curve. “Within a month, basically we had converted our services over to AKS and had everything running.”

> AKS also allowed the team to iterate much faster during stress testing, a mission-critical step in preparing for the big launch day. Before moving to AKS, it could take ForzaTech developers half an hour to swap out old images on their VMs and prepare them for load. In AKS, this process takes seconds. The extra speed enables the team to instantly make repairs, overcome roadblocks, and scale as needed. In addition, the team can script cluster updates ahead of time, adding flexibility and agility to the demanding stress tests.

> The Forza Horizon 5 architecture includes 17 core services that run in AKS. Among them, the leaderboard service tells users where they rank, while the user-generated content (UGC) service supports the ability for players to upload photos and other content. The auction house service enables players to sell cars to each other, and the analytics service collects game and player telemetry used in reports.

My current level of concurrency skills is not enough. Commercial software is all about scale. I need to broaden my knowledge if I really want to be worthy of a 200k/year salary. I want to be able to do all of the above on my own.

It is not actually big deal. Behind it all is message passing and I have a firm handle on that. But I still need some knowledge of the systems themselves. I do not want to keep my employers waiting. I might be willing to bullshit about my experiene to get the ball rolling, but I won't about my skills.

1:40pm. https://docs.microsoft.com/en-us/learn/paths/az-900-describe-cloud-concepts/?wt.mc_id=acom_newtoazure_webpage_azuremktg
https://docs.microsoft.com/en-us/learn/azure/

Let me watch the first thing and then I am going to check out some learning paths and register for Azure.

https://youtu.be/2vMEQ5zs1ko
Kubernetes vs. Docker: It's Not an Either/Or Question

Let me start with this thing.

https://youtu.be/2vMEQ5zs1ko?t=403
> The applications can now speak to each other just by using the service names that are laid out in Kubernetes.

Ok, I get it. Kubernettes works with Docker to make scaling easier. AWS makes it sound like an either or thing.

https://docs.microsoft.com/en-us/learn/paths/az-900-describe-cloud-concepts/?wt.mc_id=acom_newtoazure_webpage_azuremktg

Let me watch the stuff here.

2:10pm. This is so boring. Let me just try registering on Azure. A month will be long enough either way even if I miss a few days.

2:15pm. Nope, it also wants my card number. I am stuck for now.

2:20pm. https://docs.microsoft.com/en-us/learn/modules/intro-to-azure-fundamentals/tour-of-azure-services

Who is going to learn all this trivial. Some people have those kinds of brains suitable for tackling quiz shows, but I do not.

> The Azure free student account is an offer for students that gives $100 credit and free developer tools. Also, you can sign up without a credit card.

Let me try this thing.

> Use your university or school email to sign up and renew each year you're a student

Nope, I don't have this and I need it as expected.

https://docs.microsoft.com/en-us/learn/modules/intro-to-azure-fundamentals/case-study-introduction

Doing something like making a website would be interesting at this point as it would serve as an icebreaker. Right now I feel like I am slamming behind a glass screen.

https://docs.microsoft.com/hr-hr/learn/modules/azure-architecture-fundamentals/regions-availability-zones

Already covered this concept in AWS.

2:40pm. For the next hour before my dad gets back I am stuck, so I might as well take a break. Right now my brain is blocked up. I am at my limit how much raw text I can ingest. I need to move to digesting instead before I can consume more.

You know what, let me take a bath here, so I do not have to later. I'll commit here and resume when I got the AWS account operational. My goal for the day is to at least get the AWS Lambda to run.

In addition to everything else, doing something like a website that fetches from a database and displays it would not be bad. Some icebreaking there to get into the cloud grove would be good. I should aim to get all of the weaknesses as a programmer out of the way in one fell swoop."