"4:05pm. Hygiene - check.
4:20pm. Let me rant a bit since I have time. It does not seem like I'll be doing any programming today.

I am thinking back to the AWS videos. Remember when they talked about the Route 53, which is AWS's DNS service? It feels like I am closer to bringing it together. I have no exp in web services, and I've been really wondering what I would do if I had to do this all from scratch, but AWS really does do the hard stuff of running a server. The same goes for the other cloud providers. As long as I play in that sandbox, I could become a web expert without having to understand the intricacies of networking at the low level.

All I really need to understand are the fundamentals of concurrency in other words.

Even if did a distributed large NN model, there is only one way to transfer data between instances and that is using messages. Messages are the universal standard no matter what protocol you use.

My confidence is really high now that I've stepped into this.

As trite as the AWS intro was, it mostly presented me with solutions to the problems I'd have. The cloud gives me everything I need to scale. As an individual human I am not going to get a better offer than it.

An AI that needs to run its mind on a cluster 24/7 with full utilization could realize benefits by managing its own hardware, but I do not have the mental resources of Skynet.

It matters little. If Skynet could make 100$/h, then it won't matter much to it whether it is spending 1$/h or 0.01$/h to run itself. It should be able to cover the cloud costs either way. So I should not worry about this either.

This is a counterargument towards my previous position. It felt right that I should be able to get my own AI chip, but feelings are just that.

I was in fact on the wrong path, so it is no wonder that I've failed.

5:05pm. I've had a talk with my dad, and he suggested I pass in the number of my account as my card. Somehow that did not even occur to me. I financially illiterate by this point. How do I even log into my netbanking service? Let me check out the journal from last year. I think I revisited that then.

I can't find anything. Maybe it was even further than a year back.

Now what is my username at the bank?

I remember going to a bank years back and doing something there. I think that was to renew my card. Maybe it was to open the Spiral Collective thing?

5:20pm. I am confused how this the login card is supposed to be used. I found my username I think, but what about the password? ...I threw it into the thash 15m ago along with the 2018 annual report.

I got it back.

> Niste korisnik NetBanking aplikacije. Uslugu možete ugovoriti u poslovnici Banke. (Br.poruke: 15012)

It is outright telling me I am not a Netbanking user. The reason this is happening is because netbanking is probably for business users now, but personal accounts are some weirdly named mobile service called Georges. I can't log in to see whan my forex account number is. But I think it does need a card for this, not the account number, right?

5:35pm. This is going to be hard. My father is being stubborn about letting me borrow his card so I can register for the free tier, and now I am in a humiliating situation of having to needle him about it.

I am thinking of trying to pass off the expired card. He mentioned something about it being mailed by the bank, but that I don't remember that ever happening. He also mentioned that new card I'll get would have the same number. But if it turns out to be inactive, I'd end up wasting the entire day tomorrow.

...No, I forgot. It is asking me for the expiration date outright. I have no idea what would be for the newer card as I do not have it.

5:50pm. I give up. I guess I'll have to find something else to do other than AWS tomorrow as it is Sunday. I am 35 years old, so I am not going to pester my parents until things go my way. Doing that would be too degrading.

> We will not charge you for usage below AWS Free Tier limits. We may temporarily hold up to $1 USD (or an equivalent amount in local currency) as a pending transaction for 3-5 days to verify your identity.

Maybe this whole thing is for nothing - I could get outright rejected if I give another person's card. This could just lead to further trouble down the road.

Let me avoid this. He is basically doing me a favor here, even if he is playing it safe merely out of instinct.

5:55pm. I think for the first time in a while, I have an excuse to take a vacation. I'll visit the bank on Monday and hope they have the card waiting for me in their vault, but if they don't and require me to wait until it can be reissued, so be it.

6:05pm. Miserable, just miserable. The next week looks dark indeed.

C backend TODO:
Tuple/Layout/Closure env size ordering

I'll take the opportunity to take care of this particular issue tomorrow then, since I have the time. I could also go back to work on that skeleton in Blender while I am at it. For me, it has become a matter of pride to pursue something. If playing games would give me power, I'd play those.

It is not like I've given up on art. I just decided that I wanted money after all. Just because I want money more does not mean I do not want to do Heaven's Key. It is merely a matter of prioritization.

What I want I want most right now is to level up my cloud skills to the point I could tackle any challenge in the wild whatsoever. It is a must have skill for working at startups, and I am not going to get a support role, but be right in the thick of it.

It is important to me. It is like food, when you have it, you do not think about it. But if you don't, it is the center of all your concerns."