"8:50am. It is go-to-the-bank day. Right now I have some time before I leave, I have to wait for the mobile to charge up.
...I just realized - my personal card expired last year as well.

9:15am. It is time to go.

10:45am. I am back. Let me give Netbanking a try again.

...It is not working. Let me call the hotline.

...Wait, the paper I got says George. I've been trying to log into the wrong service. Let me try again.

11am. Everything is in order now. I changed the password for the login card and logged in just fine. Now I can see my various bank balances. I've been told that those separate accounds will all get unified into one at the start of next year.

That is fine and dandy, but now I have another problem.

I have wait until the bank mails me the new cards. That could easily take a week or longer. I have only myself to blame for not doing it ahead of time last week while I was working on the C backend.

Actually the thought to do this did arise, but I wanted to comparmentalize the tasks and focus on the present instead of burdening myself with different things.

10:10am. Stored away the old papers.

...Ok. I suppose this is one way to use the morning session. It simply needed to be done. Once I get the card and register for AWS, I'll have plenty of money to play around with it.

I have to think how I want to spend my time over the next week or so, until I get access to it.

11:20am. Right now I need to chill a bit to clear my head from the trip. Let make a list:

* Update VS to 2022.
* Check out if there is a new F# version out.

Let me set it to download.

11:25am. I am thinking which workloads to check.

Since I have a week of time in which to do random work, I am thinking of getting familiar with ASP.NET and general networking. I only dabbled in this back in 2020 when I studied concurrency seriously.

Honestly, I am ashamed of how little I learned regarding the big frameworks. ASP.NET seemed like a giant mess, I ended up using ZeroMQ instead for the language server. Because it was messy I did not even figure out how it could be used for web apps.

Checking ASP.NET only increases my download size by 0.01Gb, so I'll get it.

But rather than focus on ASP.NET I think I want some general networking knowledge. I wonder if there is a course for that sort of thing? I am going to look for it.

Until the two account cards arrive, I guess I'll be revisiting early 2020. This time I want to get more comfortable with the web in general as opposed to just concurrency. Concurrency I **know**. Any mail?

1 accept, 1 reject.

That is 3 accepts so far, out of 9 companies that I've applied to. Not bad.

I got 2 rejects, but only one is by a HR bot like now. The other one is because I was not a good fit for the Haskell guys.

11:35am. This accept for the interview is from the DL Research Framework position that I am particularly interested in. That is even better.

11:40am. It is a good chance the the other 4 I am waiting on will be accepts. This will allow me to get some good interviewing experience. I'll go from the least interesting to most interesting positions so I can get into the groove and learn from experience as I go along. I need to think about my behavior and adjust it as I go along. Being able to get practice for the actually important stages of getting a job is a great benefit of having a high hit rate. It starts you off from a position of strength.

Now, VS 2022 has finished installing, but I can't find the shortcut for it anywhere. What the hell?

11:50am. It did make a shortcut, it is just that the search is acting weird. I should have clicked on Apps to make sure. That clears that mystery.

11:55am. I can just let those accepts sit there for a few weeks. It does not matter. I'll just apply somewhere else if the positions get closed in the meantime.

https://arxiv.org/abs/1902.03383
> Serverless cloud computing handles virtually all the system administration operations needed to make it easier for programmers to use the cloud. It provides an interface that greatly simplifies cloud programming, and represents an evolution that parallels the transition from assembly language to high-level programming languages. This paper gives a quick history of cloud computing, including an accounting of the predictions of the 2009 Berkeley View of Cloud Computing paper, explains the motivation for serverless computing, describes applications that stretch the current limits of serverless, and then lists obstacles and research opportunities required for serverless computing to fulfill its full potential. Just as the 2009 paper identified challenges for the cloud and predicted they would be addressed and that cloud use would accelerate, we predict these issues are solvable and that serverless computing will grow to dominate the future of cloud computing.

Google Autocomplete was really strong in suggesting this paper for some reason. I'll check it out later.

Let me find a cloud computing course, something that has hands on practice and is not just an overview.

https://www.youtube.com/results?search_query=cloud+computing+course

Youtube has something as expected. There are a few big ones. A 10h course I'd need a few days to get through. Just the right kind of time consumer for somebody in my current position.

Let me have breakfast here. I spent the morning well. I'll get started on the course past noon."