"7:50am. I am up.
7:55am. I think this might be a little too early to get up, but it is fine. In 3h, at 11:15am I have an interview with Neurolabs.

8:40am. Let me get started. Let me watch a Python pSql vid.

8:45am. https://www.postgresqltutorial.com/postgresql-python/

No problem, there are tutorials. Let me watch a video first. After the Neurolabs interview who knows what goal I will have. Ideally, I would study a single tech stack like ASP.NET and get a job with it, but that is not realistic. I have to gain proficiency in whatever the employer is using.

8:55am. It really is a drag to seriously try to commit to learning anything right now. I'll go to the NeuroLabs interview in 2h, and no doubt they will have another tech stack.

I do not mind commiting myself to learning something, but I do not want to commit myself to learning several things.

Hopefully, I'll get a mobile today so I can register to AWS and start studying distributed ML on it.

9:05am. https://youtu.be/M2NzvnfS-hI?t=415

I really like this. Note that he just gives the relevant data to a function and gets back the context. Constrast that with .NET where you need to do several magical things.

https://youtu.be/M2NzvnfS-hI?t=579

This way of creating a table is quite straightforward compared to doing migrations.

https://youtu.be/M2NzvnfS-hI?t=742

This is really nice. It is very easy to learn. Still, once you have EF up and running, it would be less type safe and maintainable.

https://youtu.be/M2NzvnfS-hI?t=1126

I see. This is why there is an update, instead of a get and set.

9:30am. https://youtu.be/M2NzvnfS-hI?t=1259

This is a good video. I am very at ease with this sort of programming instead of the .NET stuff.

https://youtu.be/M2NzvnfS-hI?t=1357

The with clause won't close the connection? That feels wrong. I should double check that, but nevermind.

9:40am. This is nice and clean, I do not really feel like studying it for longer. I thought it would be a huge mess like EF.

Incidentally, what I feel like studying more today is EF.

9:45am. Let me just go through the Ishizoku raw and then I'll get started on that. I actually haven't read the mange, so I do not mean to imply that I am catching up. I will pick it up.

10:15am. Dad found one of his old mobiles for me. Now I can get to work on doing what I really want. Also the weather is turning sour and now a cool breeze is blowing through my room. So nice after the sauna like temps of the past week.

Ok, first of all, let me see if I can find a way to create an mToken in the web George app. Then I'll be able to activate the mobile one. This is the most ideal way to spend the morning.

10:20am. Hmmm, I checked my mail and I see I got a reject at Kalepa. I wonder why? Who knows. I didn't expect to get a reject like this, but after some behaviral interview.

I guess I'll move on to the next one.

10:25am. I am distracted. Focus me.

10:30am. No, it is no good. I'll have to make a trip to the bank to get an mToken so I can actiate the mobile George.

10:40am. It will have to be after the interview which is in 35m.

10:45am. What a drag. The only thing I am glad is that HR did not waste my time with an initial HR problem. I feel it is a red flag that the interviewing pipeline has so many behavioral tests. The company feels controling.

10:50am. At this rate, I am going to have to expand my reach. I'll look into AI jobs...

https://ai-jobs.net/

Something like this.

Otherwise I'll have to expand my vision to full stack dev jobs as well as lowering my salary range. Either way is fine. The goal of getting a job is to make money for myself, not to have fun. After 3 years of that I can just transition to trading.

10:50am. Let me read the EF framework tutorial for 10m.

10:55am. Oh wow, there is really huge thunder outside. What do I do now? I might have to reschedule?

11am. I'll reschedule.

12:35pm. I am back. There was a crack of lightning, but it passed almost immediatelly. I could have done the interview. Nevermind that, what is important is that I went to the bank and got that mToken. Let me have breakfast and I'll try registering on AWS again. It might not be George, but simply the mobile being old that was the problem.

1pm. Done with breakfast. Let me try registering on AWS.

FFFFFFFFFFFFFFFFFFFF---Nothing is showing up!

What the hell?

> Što je VISA Secure?
> VISA Secure je besplatna usluga sigurne kupnje debitnim karticama na internetu. Prilikom potvrde online kupnje, korisnik se autentificira, odnosno banci potvrđuje identitet prijavom u mToken biometrijom (PIN telefona, otisak prsta ili prepoznavanje lica na pametnom telefonu).

So do I need to use this Visa secure service?

1:10pm. My bank is really fucking with me here. Let me try calling it again. I think the teller yesterday gave me bad advice. What was the number again?

1:15pm. Everybody is busy, I can't get in touch.

Let me send an email. This time I'll attach a screenshot.

1:20pm. Done. Let me close everything. The thunder is making me nervous."