"///
I did a C backend for [Spiral](https://github.com/mrakgr/The-Spiral-Language). I got tired of art, and wanted to do programming so I finally decided to close that regret from last year. It is going to serve as a prototype for possible AI chip backends. The cool thing about it is that its ref counting is tail call optimization compatible. Now I just have to wait until I can get my hands on the chips.

An interesting story from this month is that I decided to try lying on my resume - I made one up with just generic Microsoft, Google and Tesla experience and I tried applying to 9 startups at AngelList. My actual one has like a 10% hit rate when I passed it around last year to various remote positions. Using this, I got 5 accepts. Out of that 1 got rejected when the recruiter actually looked at it and wrote a thoughtful reply saying their product does not need ML, but would be interested a year down the road if I am still on the market, 1 was a weird reject where the recruiter's computer supposedly broke down and I haven't heard from her since for the sake of rescheduling, and in the end I went to 3 interviews. I never got past those to the actual tech interviews which is a pity, since I actually have high odds of beating those, but the actual intro calls with the recruiters that have zero programming are probably impossible for me right now. This is a heavy blow to me, and I've been sorting out my feelings.

The biggest problem for me is the mind games they play to gauge my enthusiasm for joining their dinky startup. "How do you feel about working at a startup compared to a big company like Google?" "I don't have any preference." Note that she asked that at two separate points. Saying I didn't care was a big mistake. At the last place I applied to the recruiter who was really good and towards the end, played a trick on me by playing a purposely, glitchy video to demonstrate the simulated environment the company made. As I am watching she comments: "The video is glitchy." "I don't mind." Like the other recruiter she repeated the comment two times, and my response was the same. I should have feigned disappointment, but even had I realized the trap I got up at 8am after a restless night and the interview was literally at 9:30pm. It was hell. Had I known the nature of these interviews, I never would have scheduled it so late. Another problem is the general impression I give off. I am constantly bottling down my nervousness, and to make matters worse, after operating solo for so long I've forgotten how to interact with actual people. A scam artist could clear this stage easily, but for me it is very hard.

The fact that I am nervous is embarrassing for me, and a part of the reason is that I am unsure of whether I can actually do these jobs. It is an irrational impulse, much like the impulse to check at night whether I've let the tap water on by accident. I can't really prove to myself that I could do the jobs despite my skill, and the thought starts spreading like cancer. I am just too green to take any of these jobs and the problem is not my programming skill. Even if I had 30 years of experience doing my own projects I'd still have this nagging doubt, that even if I can make anything for myself whether I could do the same for others.

At this point to get rid of my weaknesses, I am rolling the idea of freelancing in my mind. I just remembered that it is a thing today. I've looked at it in the past, but the transactions were so low that I immediately put it out of mind and forgot about it as an avenue for work, but if it is a matter of pride to prove to myself that I can do it, then I could justify attempting it. In the past I've always had my own projects so I couldn't bother to waste time with something like freelancing, but now it has some value to me. Similarly to picking up fat chicks before moving on to pretty ones, it will boost my confidence when dealing with recruiters in the future and give me some badly needed social skills. The projects I do on a freelance site will be experience I can bluff about as being a part of some company in the future, which will make my lies a lot more realistic that the generic popular product kind like 'I worked on vision data pipeline for the self driving car at Tesla' that I put in the resume. I actually can't remember half of what I put on the resume. It will get me familiar with real world uses of programming, most people don't do it for the same reasons as me so my values are out of sync with the market. Even a junior with 3 months of experience could pose better as a seasoned pro than me.

Freelancing is like the micro stakes in online poker. No good players play at those stakes so my actual resume which the recruiters would chuck into the wastebasket, might actually be good there and the general story as to why I am there might even make sense to those willing to take the risk of giving me an assignment, unlike the senior-Googler-applying-to-the-small-startup one.

Jobs - even highly paid ones, are quite unattractive to me because they are more than just work. They are life paths, and they involve doing work that I do not see as being meaningful.

10k dollars is not much, but if you think about it I could use that to upgrade my rig and buy the AI chips when they come out, so what more do I need? What would be meaningful work for me would be programming novel hardware. When I threw in the towel on poker botting last year, I blamed it on the hardware, so getting my hands on AI chips became an obsession for me. You could probably talk me into programming those things for you regardless of the salary. I want to ask those [machines a question](https://github.com/mrakgr/The-Spiral-Language/blob/fcca1f4cdac94ca6e472b34a560404e12079d08c/A%20Question%20for%20the%20Machine.pdf) of how to learn. When they become more widespread in a few years, programming them will be its own specialization, so I’ll be able to get a job without internal conflicts then.

Anyway, I do not feel like freelancing now, but doing some sci-fi writing instead. I've been learning art for the sake of illustrating that story, and I want to forget everything for a while and just write like I've been wanting to. Until the era when I can program my mind arrives, I'll have to maintain the skill with some cheap side projects. I am not really expecting to make anything from writing, though if that happened it would be very cool. I haven't written anything yet, I'll make an announcement on [my Twitter](https://twitter.com/Ghostlike) when I publish the initial chapters.

///

I meant to write `for you` in that line."