"3:55pm. https://techcrunch.com/sponsor/g-core-labs/accelerate-ai-ml-with-the-new-ipu-based-cloud-platform-by-g-core-labs-graphcore/
I am going through the news trying to get a sense of how old Graphcore's cloud offering is.

`<meta name="sailthru.date" content="2022-07-14 11:54:45" />`

There isn't a published date in the article, but I found this in the source.

`"dateCreated":"2022-07-14T18:54:45Z","datePublished":"2022-07-14T18:54:45Z","dateModified":"2022-07-15T17:40:17Z"`

So it was only published two weeks ago.

Thanks to this, I can conclude that Graphcore's cloud is new.

https://cirrascale.com/graphcloud.php

Holy shit these prices are huge.

3k a week? 20$/h. It is like I am made of gold.

> * Cirrascale Cloud Services does not provide servers by the hour. The "hourly equivalent" price is shown as a courtesy for comparison against vendors like AWS, Google Cloud, and Microsoft Azure.

Don't tell me I have to rent it for something like an entire week?

https://www.enterpriseai.news/2021/01/29/graphcore-ipu-gets-a-public-cloud-boost/

Actually, it was from a while ago. So the Graphcore cloud has been there for a while.

4:05pm. Anyway, the plan I've been toying in my mind is to make a Graphcore & Tenstorrent backends and run a little game on them. Well, not a little, but Texas Holdem. I could also make a Python backend. I could then use that to advertize and seek out sponsors. 'Look how easy it is to program these chips in Spiral. This is the power of functional programming with partial evaluation.'

It would be a lot of work, but it would be fun to do. But the above prices are simply prohibitive.

Is my quest going to end here? At this rate, I am simply going to get priced out of the Singularity. Isn't there anything I can do here?

https://www.youtube.com/results?search_query=interview+practcie

I am going to put some effort into studying this just in case.

https://www.youtube.com/watch?v=srw4r3htm4U
Job Interview Simulation and Training - Mock Interview

Let me watch this.

...Actually nevermind that for now. I do not feel like simulating interviews.

4:20pm. Maybe the cloud platform by G-Core is something other than Cirrascale and it might be cheaper.

4:25pm. There is no way around it. After I've gotten a good night's rest, I will go into the TensTorrent discord and do the following.

* Ask whether I could take a look at the docs despite not being a customer yet. Note that I am interested in how their chip could be programmed at a low level. Explain the `Question To A Machine` proposal.
* Ask whether they'd be interested in a functional programming language such as Spiral to replace C++ and show it to them. Touch upon the benefits it has. Ask whether they have any customers who are using it for purposed other than purely deep learning and are interested in the simulation capabilities of the chip.
* Ask whether they'd be interested in hiring a functional PL expert such as myself.
* Tell them I'd be interested in implementing a Spiral backend for the TensTorrent chip as well as a demo to showcase it. It will be the Texas Holdem game. I'll try to implement the training process as well and test how many hands per second that gets me.

4:50pm. Hmmmm...I'll try that. But if that fails, I am not sure what I'll want to do. Writing after all?

My interest in regular programming tasks is nil. It is easy to come up with a plan as if it were a game:

* Grind freelancing.
* Practice interviewing.
* Apply to jobs.

And so on. This is how I'd guide a player character here. But subjectively, I am always operating at the limit of my will.

5pm. I want to write, and I want to program, but I can't do both.

Maybe after finishing that C backend, I've already peaked as a programmer? I feel really lame to end the journey like this - by expending my willpower and retreating into the land of dreams. Considering I want to cause the Singularity, you'd imagine such a person would continue pushing forward forever until he gets his way, instead of accepting the encroaching demise.

5:10pm. In this situation what I really want to do is grind AI chip related apps. If I wanted to be a regular programmer, I'd make vanilla apps. The kinds of programs I am interested in are RL agents and simulations on AI chips. But I can't think of a way to turn profit on those?

Short term financial trading? I do not have the capital to start that. And even so I am not interested in short term trading, but longer term.

There are just so many things you can do in this world, it is maddening. It is not like in a static game where you are invited to do all the side quests.

5:15pm. It is not actually such a bad thing to try writing for 6 months or a year. If try it and get no success, I can take it as a sign that there is no point in me being a writter. That would make regular programming jobs attractive.

Right now I am 35. If I write for a year or two, I'll still have two decades to apply myself as a human programmer. Maybe I could keep going till I am 70 or even 80 even.

I think a decade of programming would be enough to wipe away my shame. I'll get a junior position somewhere, even locally if need be, and work my way up from there.

5:20pm. It depends whether I can get to 1k/month as a writter which is the average yearly wage in Croatia.

It is a risk, everything is a risk.

Compared to 2014, my art skill really is better so I won't have trouble with the covers again.

5:40pm. It is not the end of my life, but it is the end of my path a solo dev. If I want resources and money I need to turn pro.

I'll forget the pride of 6.5 years of struggle on my own and start from the ground up. It will be like an adult going to kindergarden, but I have to start somewhere and that somewhere will have to be at the bottom whether that means freelancing or junior wage positions. I won't aim for senior right off the bat. I can laugh at those midwit senior Googlers who don't understand linked lists, but they have a point that this is not used professionally, and I do not know what is. I have to acknowledge this weakness of mine. They might not have programming skills, but they do have what I don't.

5:45pm. Either I step on the path of a pro visual novelist or a pro developer.

I am going to decide that soon. That trip to the Discord will decide it for me.

I'll step on the pro programmer path working on AI chips or I will write stories about the Singularity.

If that Discord is just one guy doing tech support then that is fine, I guess I'll go the novelist route. But if it turns out there are active conversations about the use of TensTorrent's chips by a community, I will have a possibility of making something of Spiral.

TensTorrent is probably the only choice. Graphcore's prices are extremely high, I can't imagine it being used for anything but deep learning.

6:10pm. I think the scenario is highly likely that I'll go the novelist route. My luck when it comes to programming was always bad. Nothing good has happened to me in programming.

6:15pm. Let me close here. Tomorrow I will do a write up for the TensTorrent discord and take the plunge. After that I'll take a break, and most likely begin writing for real.

Sigh, I remember when I stopped writing Simulacrum in 2014. I was so sad to leave it in favor of programming. I'll be similarly sad to leave programming behind too. But I'll be back at some point if I am able. When I was 27, my succeesses during my schooldays and my talent summoned me back to it after a decade and half of absence.

Maybe Spiral will be calling me back at some point too."