"8:10am. It is time to give this a try. Any mail? Not anything worth mentioning.
Before I go in, let me do the pitch.

///

Hello, I am Marko Grdinić. Out of the 6.5 years that I've been working on my own projects as a hobby programmer, 3 years of full time work went into creating the language called Spiral. I started work on it in 2017 specifically with the intention of using it on future AI chips. I had a lot of difficulty creating a GPU-based ML library in F#, so I wanted a language that would be easy to use, highly efficient - enough to compile GPU kernels on it, as well being capable of seamless language interop. I haven't been interested in GPUs for a while now, and I really want the chance to prove that my vision was true and that the tool is suitable for something like Tenstorrent's chips.

https://github.com/mrakgr/The-Spiral-Language

It is a functional language similar to F#, except with some extensions that greatly increase the domain of what it can do compared to regular languages. If any in this community are interested in the simulation capabilities of Tenstorrent's chips rather than just using their functionalities from PyTorch, you'll find the language very useful. For writing kernel code directly, it will be much easier to use than C++. You can check out the language on the VS Code marketplace. Recently I've done a ref counted C backend for it.

My proposal is to implement a Tenstorrent chip backend so the language can be used with Tenstorrent chips. My request is that I either be given access to the cloud and free credits for it, or that you mail me one of the Tenstorrent cards to my address in Croatia for free. Once I have that I'll implement the backend, and do some programs as demos to show how it can be used.

How does that sound to all of you here?

///

This is good. Now where is the Discord link?

https://docs.tenstorrent.com/tenstorrent/learn-more/support

8:40am. It is asking me for my mobile now. Good thing I got that dealt with and actually have one now. It would have been a pain to have to go through dad's just to get the registration code. I had to do that several times in the past.

8:45am. Ok, I am in. Let me post the above. I might just get banned for self promoting, but so be it.

8:50am. The channel is quite slow. But this way I can guarantee it will be seen and thought about by the relevant engineers at Tenstorrent.

8:55am. Well, I'll check it every hour or so. I slept well during the night today. Yesterday I was obsessed about what to do today so I did not have much to think about what to write.

I've been reflecting on my mistake.

In order for me to make money via programming, I need to get over my inhibition of doing favors to random people.

Solo dev and pro dev two different paths. Even if I wanted those 100-200k remote jobs, they are absolutely impossible for me until I've made the determination to step on the pro path. In order to do that, I'd need to either do freelancing or get a junior position for 3-6 months. I do not feel like going down that route. With the exception of satisfying my obsession towards AI chips in this thread here, I am not interested in stepping on the pro path.

9:15am. Let me close the tab. I'll check out the Discord later in the day. It is very slow, let me just leave it out of mind for now.

Doing this was the first good move that I did since last year apart from doing the C backend. The companies are too difficult approach. I wish I had thought of looking whether they had a Discord server earlier. But hasn't been more than 7 months since they opened it, and people only started joining 5 months ago. So I did not waste too much time.

Right now, Tenstorrent is my best bet for making Spiral useful on these devices. Other companies do not have the vision to see the benefits of serving the little people. While big customers will make them the big money, it is the individual programmers who will write the libraries that will make their devices really valuable to use. So it is good to spread cheap version of their AI chip so that people with small budgets can play with them.

Companies like Graphcore and Groq will have a strategic problem years down the road when Tenstorrent will have OS libraries made for it by the community while they are stuck paying programmers to do the same. This is what I meant by Tenstorrent having the strongest vision compared to Graphcore and Groq in the `A Question for the Machine` pdf. Incidentally, this is also why you wouldn't want to be in inference-only business. The more general purpose the device is, the better the results will be in the long run for the company.

To be honest, I expected more from the 20s. There are many startups in the field, but only a few are interesting. I thought I might be able to have multiple sponsors for Spiral easily once the wave hits, but I'll be lucky if I manage even 1 at this rate.

9:40am. Let me step away from the screen. I've spend quite a lot of time in the last few days in bed, and it seems today won't be an exception.

I'll assume that my offer won't get accepted and start focusing my will to step on the pro novelist path.

2014 when I wrote the first Simulacrum arcs was special. I didn't really care about my audience too much, I just wanted to write the story as a way of declaring victory. To say I understand what the Singularity is about. But this time, at least I will make sure to post it on a site like Royal Road where most people can see it. I'll open a Patreon with advance chapters. An important part of my writing journey is to really gauge interest in this kind of story.

This kind of story is the only thing I want to write. I woudn't be able to write anything else actually. Maybe that has a parallel with my programming.

You'd think that a pro writter would be able to write any kind of story, but I want only one kind. So maybe I am not suitable to be a pro writter either.

But it is fine if I am no good as a pro. My programming is good and my writing will be good. What else can I ask for?"